restructured payment
