+++ src/registrationSystem/utn_pay/payment_models/abstract_payment_base.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class AbstractPayment(ABC):
+    @abstractmethod
+    def get_order_rows_and_cost(self, team_leader):
+        pass
